Enhance Sprite Editor with New Features and Refactor Codebase

+++ src/editor/actions/delete_action.h
@@ -15,5 +15,5 @@ public:
 protected:
     std::shared_ptr<moth_ui::Node> m_deletedNode;
     std::shared_ptr<moth_ui::Group> m_parentNode;
-    size_t m_originalIndex = 0;
+    int m_originalIndex = 0;
 };
